Update README.md
so basically using these lines of codes you can use your own chatgpt for free and modify it. in this sample you can ask it the prices of various cryptocurrencies. keep in mind that in the first blank you have to type your prompt then in the second blank you have to type your email address, and the last but not the least, you have to choose the language which the response has to be in.